LayoutTests/media/modern-media-controls/scrubber/scrubber.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>Scrubber</code> class.");

const scrubber = new Scrubber;

shouldBeTrue('scrubber.element.classList.contains("scrubber")');

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
